refactor: component filter
